Rename dpos state files
